Working on study and assay details update.
