+++ client/src/pages/assayDetails/AssayTimelineTab.js
@@ -15,7 +15,7 @@
  */
 
 import React, {useEffect, useState} from "react";
-import {Col, Row} from 'react-bootstrap'
+import {Card, Col, Row} from 'react-bootstrap'
 import {ActivityStream} from "../../common/activity";
 import {CardLoadingMessage} from "../../common/loading";
 import {DismissableAlert} from "../../common/errors";
@@ -59,13 +59,15 @@ const AssayTimelineTab = props => {
   }
 
   return (
-      <div className="timeline-tab">
-        <Row>
-          <Col sm={12}>
-            {content}
-          </Col>
-        </Row>
-      </div>
+      <Card>
+        <Card.Body>
+          <Row>
+            <Col sm={12}>
+              {content}
+            </Col>
+          </Row>
+        </Card.Body>
+      </Card>
   )
 
 }
